Make ddtxn go-gettable.
This makes a number of changes:

* Changes all `ddtxn/*` imports into the full `github.com/narula/ddtxn` paths.
* Moves all benchmarks into their own directories. This allows the use of `go install` to build and install the binaries.
* Changes `bm.py` to expect the benchmark binaries to be in `$GOPATH/bin/`.
* Freshens the README a bit.
