+++ apps/big.go
@@ -1,8 +1,9 @@
 package apps
 
 import (
-	"ddtxn"
 	"flag"
+
+	"github.com/narula/ddtxn"
 )
 
 var incr = flag.Bool("incr", true, "Do incr or RW workload")
