+++ benchmarks/big/big.go
@@ -1,10 +1,6 @@
 package main
 
 import (
-	"ddtxn"
-	"ddtxn/apps"
-	"ddtxn/dlog"
-	"ddtxn/prof"
 	"flag"
 	"fmt"
 	"log"
@@ -13,6 +9,11 @@ import (
 	"runtime"
 	"sync"
 	"time"
+
+	"github.com/narula/ddtxn"
+	"github.com/narula/ddtxn/apps"
+	"github.com/narula/ddtxn/dlog"
+	"github.com/narula/ddtxn/prof"
 )
 
 var nprocs = flag.Int("nprocs", 2, "GOMAXPROCS default 2")
