+++ benchmarks/buy/buy.go
@@ -2,10 +2,6 @@ package main
 
 import (
 	"container/heap"
-	"ddtxn"
-	"ddtxn/apps"
-	"ddtxn/dlog"
-	"ddtxn/prof"
 	"flag"
 	"fmt"
 	"log"
@@ -15,6 +11,11 @@ import (
 	"sync"
 	"sync/atomic"
 	"time"
+
+	"github.com/narula/ddtxn"
+	"github.com/narula/ddtxn/apps"
+	"github.com/narula/ddtxn/dlog"
+	"github.com/narula/ddtxn/prof"
 )
 
 var nprocs = flag.Int("nprocs", 2, "GOMAXPROCS default 2")
