+++ execute.go
@@ -1,10 +1,11 @@
 package ddtxn
 
 import (
-	"ddtxn/dlog"
 	"flag"
 	"log"
 	"math/rand"
+
+	"github.com/narula/ddtxn/dlog"
 )
 
 var SampleRate = flag.Int64("sr", 500, "Sample every sr transactions\n")
