+++ util.go
@@ -2,7 +2,6 @@ package ddtxn
 
 import (
 	crand "crypto/rand"
-	"ddtxn/dlog"
 	"fmt"
 	"log"
 	"math"
@@ -10,6 +9,8 @@ import (
 	"os/exec"
 	"strings"
 	"time"
+
+	"github.com/narula/ddtxn/dlog"
 )
 
 type Exp2 struct {
